requirements.txt
# Supply Chain Management API - Dependencies
# ===========================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
python-dateutil==2.8.2

# Testing (Optional)
pytest==7.4.yncio==0.4
pytest-as23.3
httpx==0.26.0 (Optional)
black==

# Development24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
